♻️ Refactor run approval+denial to shared logic
